# docs/swagger.yaml
basePath: /v3/ord
definitions:
  models.Coverage:
    properties:
      amount:
        example: 60
        type: integer
      emission_id:
        format: uuid
        type: string
      order_id:
        format: uuid
        type: string
    required:
    - amount
    - emission_id
    - order_id
    type: object
  models.Emission:
    properties:
      base_currency:
        example: EUR
        type: string
      base_currency_amount:
        description: BaseCurrencyAmount is summary fiat amount covered by this emission
        example: 100
        type: integer
      coverage:
        description: Coverage is an array of pairs 'order - coverage amount' that says how much the order is covered by this emission
        items:
          $ref: '#/definitions/models.Coverage'
        type: array
      created_at:
        example: 1598884059
        type: integer
      details:
        description: Details is detailed information about emission in raw json format
        example: '{''field1'':''value1'', ''field2'':''value2''}'
        format: json
        type: string
      gold_amount:
        example: 30
        type: integer
      gold_unit:
        example: grams
        type: string
      id:
        format: uuid
        type: string
      num:
        description: Num is a serial number of the emission in the database
        example: 3
        type: integer
      rate_base_to_unit:
        example: 12.5
        type: number
      reference:
        description: Reference is any additional (unnecessary) information about emission
        example: any text
        type: string
      type:
        enum:
        - deposit
        - withdraw
        type: string
      updated_at:
        example: 1598884069
        type: integer
    required:
    - base_currency
    - base_currency_amount
    - coverage
    - details
    - gold_amount
    - gold_unit
    - rate_base_to_unit
    - type
    type: object
...
  order-manager_pkg_api_v3_delivery_rest_handlers.ResponseDataCreate:
    properties:
      emission_id:
        format: uuid
        type: string
    type: object
...
  render.BodyError:
    properties:
      message:
        example: something went wrong
        type: string
      status:
        default: error
        type: string
    type: object
  render.BodySuccess:
    properties:
      data:
        type: object
      status:
        default: success
        type: string
    type: object
info:
  contact:
    name: Some Cool Company
  description: Endpoints documentation.
  license:
    name: All rights reserve.
  title: Order-Manager REST API
  version: "1.0"
paths:
  /adm/emission:
    post:
      consumes:
      - application/json
      parameters:
      - description: Token retrieved from admin-api by /auth endpoint.
        in: header
        name: Authorization
        required: true
        type: string
      - description: Emission content
        in: body
        name: emission
        required: true
        schema:
          $ref: '#/definitions/models.Emission'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/render.BodySuccess'
            - properties:
                data:
                  $ref: '#/definitions/order-manager_pkg_api_v3_delivery_rest_handlers.ResponseDataCreate'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/render.BodyError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/render.BodyError'
      summary: Create
      tags:
      - Emission
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"